refactor: move UI elements to their own folder
